requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.6
django-crispy-forms==2.1
django-js-asset==2.2.0
django-tinymce==4.0.0
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
